refactor : category 도메인 분리
