+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/category/dto/CategoryIdResponseDto.java
@@ -1,4 +1,4 @@
-package com.sparta.blackwhitedeliverydriver.dto;
+package com.sparta.blackwhitedeliverydriver.category.dto;
 
 import java.util.UUID;
 import lombok.AllArgsConstructor;
