+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/category/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.blackwhitedeliverydriver.repository;
+package com.sparta.blackwhitedeliverydriver.category.repository;
 
-import com.sparta.blackwhitedeliverydriver.entity.Category;
+import com.sparta.blackwhitedeliverydriver.category.entity.Category;
 import java.util.Optional;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
